+++ headless-Checkout/client/react/headless-checkout/src/pages/headless-checkout-options/HeadlessCheckoutOptions.jsx
@@ -4,7 +4,7 @@ export default function HeadlessCheckoutOptions() {
     return (
         <div className="container flex flex-column gap-20">
             <Link to='/make-payment' className="btn btn-1">MAKE PAYMENT</Link>
-            <Link to='/save-payment-method/payment-method-options' className="btn btn-1">SAVE A PAYMENT METHOD</Link>
+            <Link to='/save-payment-method' className="btn btn-1">SAVE A PAYMENT METHOD</Link>
             <Link to='/pay-with-saved-payment-method' className="btn btn-1">PAY WITH SAVED PAYMENT METHOD</Link>
             <Link to='/validate-fields' className="btn btn-1">VALIDATE FIELDS</Link>
             <Link to='/get-card-info' className="btn btn-1">GET CARD INFO</Link>
